fixed all warnings

+++ src/bumpalloc.cpp
@@ -7,7 +7,7 @@ static uint8_t *begin = (uint8_t *)&_malloc_begin;
 static uint8_t *end = (uint8_t *)&_bootloader_end;
 
 void *bump_malloc(size_t n) {
-    if ((end - begin) < n) {
+    if ((size_t)(end - begin) < n) {
         return nullptr;
     }
     uint8_t *ret = begin;
